Member model corretion accordint to Anton's first revision
