+++ src/core/services/shift_service.py
@@ -60,7 +60,6 @@ class ShiftService:
 
         update_shift_dict = {"started_at": datetime.now().date(), "status": Shift.Status.STARTED.value}
         updated_shift = await self.__shift_repository.update(id=id, instance=Shift(**update_shift_dict))
-        await self.__member_repository.create_members(id)
         await self.__user_task_service.distribute_tasks_on_shift(id)
         return updated_shift  # noqa: R504
